+++ R/class_options.R
@@ -384,7 +384,7 @@ options_class <- R6::R6Class(
     },
     validate_error = function(error) {
       deprecate_error_workspace(error)
-      tar_assert_flag(error, c("stop", "continue", "workspace"))
+      tar_assert_flag(error, c("stop", "continue", "abridge", "workspace"))
     },
     validate_memory = function(memory) {
       tar_assert_flag(memory, c("persistent", "transient"))