+++ R/tar_target_raw.R
@@ -103,7 +103,7 @@ tar_target_raw <- function(
   )
   tar_assert_format(format)
   tar_assert_flag(iteration, c("vector", "list", "group"))
-  tar_assert_flag(error, c("stop", "continue", "workspace"))
+  tar_assert_flag(error, c("stop", "continue", "abridge", "workspace"))
   deprecate_error_workspace(error)
   tar_assert_flag(memory, c("persistent", "transient"))
   tar_assert_lgl(garbage_collection)